image: Keep track of the snapshotters
This will help to prevent situations where a snapshotter S1 has a
snapshot available, but snapshotter S2 is being used (which doesn't have
a snapshot for this image).

Later on we'll improve ensureImageExists() to skip the image pulling
because of what's being mentioned above, which would result in the
sandbox failing to start.

Signed-off-by: Wedson Almeida Filho <walmeida@microsoft.com>
Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>