Skip image pull only if the correct snapshotter has a snapshot
Let's take advantage of the newly added Snapshotters mapping for the
Image, and ensure that we'll only skip the image pulling in case the
correct snapshotter, the one used by the specific runtime handler, has
the snapshot.  Otherwise we'd end up returning an image from a different
snapshotter (the default one), resulting in the sandbox failing to
start.

Signed-off-by: Wedson Almeida Filho <walmeida@microsoft.com>
Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>